services/google/requirements.txt
Flask==3.0.0
google-api-python-client==2.107.0
google-auth-oauthlib==1.1.0
python-dateutil==2.8.2